+++ docs/CHANGELOG.md
@@ -58,6 +58,7 @@ Version 1.9.0 (not released yet)
 - Improve handling for players leaving/joining multiplayer games during votes
 - Add `/hasmap` server chat command
 - Add handling of level filename extensions with `/vote level` server chat command
+- Change spawn protection option in dedicated server config to allow it to use Invulnerability powerup
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------